door lock controls for HART shuttle lifecycle, and specifically for the doors that lead into the shuttle boarding hallway
